requirements.txt
bcrypt (3.1.7)
cffi (1.14.0)
cryptography (2.9.2)
future (0.18.2)
netmiko (3.1.0)
paramiko (2.7.1)
pip (9.0.1)
pkg-resources (0.0.0)
pyaml (20.4.0)
pycparser (2.20)
PyNaCl (1.3.0)
pyserial (3.4)
PyYAML (5.3.1)
scp (0.13.2)
setuptools (39.0.1)
six (1.14.0)
textfsm (1.1.0)
